SDK-4713: review fix
