+++ .gitignore
@@ -14,11 +14,5 @@ nbproject/
 
 # tooling
 vendor/
-!integrator/project_original/vendor/
-!general/project_original/vendor/
-!generator/project_original/vendor/
 composer.lock
-!integrator/project_original/composer.lock
-!general/project_original/composer.lock
-!generator/project_original/composer.lock
 .phpunit.result.cache
