--- src/Infrastructure/Helper/StringHelper.php
@@ -1,30 +1,0 @@
-<?php
-
-/**
- * Copyright © 2016-present Spryker Systems GmbH. All rights reserved.
- * Use of this software requires acceptance of the Evaluation License Agreement. See LICENSE file.
- */
-
-declare(strict_types=1);
-
-namespace SprykerSdk\Utils\Infrastructure\Helper;
-
-class StringHelper
-{
-    /**
-     * @param string $string
-     * @param bool $capitalizeFirstCharacter
-     *
-     * @return string
-     */
-    public static function fromDashToCamelCase(string $string, bool $capitalizeFirstCharacter = true): string
-    {
-        $str = str_replace('-', '', ucwords($string, '-'));
-
-        if (!$capitalizeFirstCharacter) {
-            $str = lcfirst($str);
-        }
-
-        return $str;
-    }
-}
